+++ showcase/src/main/java/org/patternfly/showcase/extension/FinderExtension.java
@@ -15,6 +15,8 @@
  */
 package org.patternfly.showcase.extension;
 
+import java.util.List;
+import java.util.Random;
 import java.util.function.Function;
 
 import org.jboss.elemento.Elements;
@@ -29,17 +31,28 @@ import org.patternfly.extension.finder.FinderItem;
 import org.patternfly.extension.finder.FinderItemActions;
 import org.patternfly.extension.finder.FinderItemDescription;
 import org.patternfly.extension.finder.FinderPreview;
+import org.patternfly.extension.finder.PreviewHandler;
 import org.patternfly.showcase.Snippet;
 import org.patternfly.showcase.SnippetPage;
 import org.patternfly.showcase.model.File;
 import org.patternfly.showcase.model.Files;
+import org.patternfly.showcase.model.Record;
+import org.patternfly.showcase.model.Track;
+import elemental2.promise.Promise;
 
+import static elemental2.dom.DomGlobal.setTimeout;
+import static java.util.Arrays.asList;
+import static java.util.Collections.singletonList;
+import static java.util.stream.Collectors.toList;
 import static org.jboss.elemento.Elements.a;
-import static org.jboss.elemento.Elements.removeChildrenFrom;
+import static org.jboss.elemento.Elements.div;
+import static org.jboss.elemento.Elements.img;
 import static org.patternfly.component.button.Button.button;
 import static org.patternfly.component.content.Content.content;
 import static org.patternfly.component.content.ContentType.h1;
 import static org.patternfly.component.content.ContentType.p;
+import static org.patternfly.component.emptystate.EmptyState.emptyState;
+import static org.patternfly.component.emptystate.EmptyStateBody.emptyStateBody;
 import static org.patternfly.component.icon.Icon.icon;
 import static org.patternfly.component.list.DescriptionList.descriptionList;
 import static org.patternfly.component.list.DescriptionListDescription.descriptionListDescription;
@@ -51,9 +64,14 @@ import static org.patternfly.extension.finder.FinderColumn.finderColumn;
 import static org.patternfly.extension.finder.FinderColumnActions.finderColumnActions;
 import static org.patternfly.extension.finder.FinderColumnHeader.finderColumnHeader;
 import static org.patternfly.extension.finder.FinderColumnSearch.finderColumnSearch;
+import static org.patternfly.extension.finder.FinderItem.emptyItem;
 import static org.patternfly.extension.finder.FinderItem.finderItem;
 import static org.patternfly.extension.finder.FinderPreview.finderPreview;
+import static org.patternfly.icon.IconSets.fas.ban;
+import static org.patternfly.icon.IconSets.fas.folder;
+import static org.patternfly.icon.IconSets.fas.music;
 import static org.patternfly.icon.IconSets.fas.plus;
+import static org.patternfly.icon.IconSets.fas.recordVinyl;
 import static org.patternfly.icon.IconSets.fas.redo;
 import static org.patternfly.icon.IconSets.fas.search;
 import static org.patternfly.icon.PredefinedIcon.predefinedIcon;
@@ -62,14 +80,16 @@ import static org.patternfly.showcase.ApiDoc.Type.subcomponent;
 import static org.patternfly.showcase.BuildingBlocks.mixedKebab;
 import static org.patternfly.showcase.Code.code;
 import static org.patternfly.showcase.Data.extensions;
+import static org.patternfly.showcase.model.Discography.records;
 import static org.patternfly.style.Classes.filtered;
 import static org.patternfly.style.Classes.modifier;
 import static org.patternfly.style.Classes.util;
+import static org.patternfly.style.Size.xs;
 
 @Route(value = "/extensions/finder", title = "Finder")
 public class FinderExtension extends SnippetPage {
 
-    @SuppressWarnings("unchecked")
+    @SuppressWarnings({"unchecked", "ExtractMethodRecommender"})
     public FinderExtension() {
         super(extensions.get("finder"));
 
@@ -153,7 +173,6 @@ public class FinderExtension extends SnippetPage {
                                         }))
                                 .onPreview((item, preview) -> {
                                     File file = item.get("file");
-                                    removeChildrenFrom(preview);
                                     preview.add(content(h1)
                                             .add(icon(predefinedIcon(file.icon)).inline())
                                             .add(" " + file.name));
@@ -173,6 +192,111 @@ public class FinderExtension extends SnippetPage {
                         .element()
                 // @code-end:finder-preview
         ));
+
+        addSnippet(new Snippet("finder-async", "Async items",
+                content(p).editorial().add("The finder below contains the ")
+                        .add(a("https://en.wikipedia.org/wiki/Red_Hot_Chili_Peppers#Discography", "_blank")
+                                .text("discography"))
+                        .add(" of the Red Hot Chili Peppers. The decades are added statically, the records are loaded ")
+                        .add("asynchronously (with a random delay and error). If there are no records in a decade, an ")
+                        .add("empty item is displayed. Once loaded the records won't be loaded again, until you reload them."),
+                code("finder-async"), () -> {
+            // @code-start:finder-async
+            int[][] decades = new int[][]{
+                    new int[]{1970, 1979},
+                    new int[]{1980, 1989},
+                    new int[]{1990, 1999},
+                    new int[]{2000, 2009},
+                    new int[]{2010, 2019},
+                    new int[]{2020, 2029},
+                    new int[]{2030, 2039},
+            };
+
+            PreviewHandler recordPreview = (item, preview) -> {
+                Record record = item.get("record");
+                preview.add(content(h1).text(record.title))
+                        .add(content(p).editorial().text("Released " + record.released))
+                        .add(div().add(img(record.cover)))
+                        .add(content(p)
+                                .add("More infos: ")
+                                .add(a(record.url, "_blank").text(record.url)));
+            };
+            PreviewHandler trackPreview = (item, preview) -> {
+                Track track = item.get("track");
+                preview.add(content(h1).text(track.title))
+                        .add(descriptionList().horizontal().css(util("mt-sm"))
+                                .addItem(descriptionListGroup()
+                                        .addTerm(descriptionListTerm("Title"))
+                                        .addDescription(descriptionListDescription(track.title)))
+                                .addItem(descriptionListGroup()
+                                        .addTerm(descriptionListTerm("Length"))
+                                        .addDescription(descriptionListDescription(track.length)))
+                                .run(dl -> {
+                                    if (track.writer != null) {
+                                        dl.addItem(descriptionListGroup()
+                                                .addTerm(descriptionListTerm("Writer"))
+                                                .addDescription(descriptionListDescription(String.join(", ", track.writer()))));
+                                    }
+                                }));
+            };
+
+            Function<Record, FinderItem> recordItem = record -> finderItem(Id.build(record.title))
+                    .text(record.title)
+                    .addDescription(String.valueOf(record.year))
+                    .icon(recordVinyl())
+                    .store("record", record)
+                    .nextColumn(finderColumn(Id.build(record.title))
+                            .addHeader(finderColumnHeader(record.title))
+                            .addItems(record.tracks.asList(), track ->
+                                    finderItem(Id.build(record.title, String.valueOf(track.track)))
+                                            .text(track.track + ". " + track.title)
+                                            .icon(music())
+                                            .store("track", track))
+                            .onPreview(trackPreview));
+
+            Function<int[], FinderColumn> decadeColumn = decade -> finderColumn(
+                    Id.build(String.valueOf(decade[0]), String.valueOf(decade[1])))
+                    .run(column -> column.addHeader(finderColumnHeader(decade[0] + " - " + decade[1])
+                            .addActions(finderColumnActions()
+                                    .addButton(button(redo()).plain()
+                                            .onClick((e, b) -> column.reload())))))
+                    .addItems(item -> new Promise<>((resolve, reject) -> {
+                        boolean boom = Math.random() < 0.25;
+                        int delay = new Random().nextInt(2000); // simulate remote call
+                        setTimeout(__ -> {
+                            if (boom) {
+                                reject.onInvoke("Random error");
+                            } else {
+                                List<FinderItem> items = records(r -> r.year >= decade[0] && r.year <= decade[1])
+                                        .stream()
+                                        .map(recordItem)
+                                        .collect(toList());
+                                if (items.isEmpty()) {
+                                    resolve.onInvoke(singletonList(emptyItem(emptyState()
+                                            .size(xs)
+                                            .text("No records")
+                                            .icon(ban())
+                                            .addBody(emptyStateBody().text("No records have been released in this decade")))));
+                                } else {
+                                    resolve.onInvoke(items);
+                                }
+                            }
+                        }, delay);
+                    }))
+                    .onPreview(recordPreview);
+
+            return finder().style("height: 450px;")
+                    .addItem(finderColumn("Decades")
+                            .addHeader(finderColumnHeader("Decades"))
+                            .addItems(asList(decades),
+                                    decade -> finderItem(Id.build(String.valueOf(decade[0]), String.valueOf(decade[1])))
+                                            .text(decade[0] + " - " + decade[1])
+                                            .icon(folder())
+                                            .nextColumn(decadeColumn.apply(decade))))
+                    .addPreview(finderPreview().css(util("p-md")))
+                    .element();
+            // @code-end:finder-async
+        }));
 
         startApiDocs(Finder.class);
         addApiDoc(Finder.class, component);